cocode/requirements.txt
openai>=1.5.0
open-interpreter>=0.1.17
textual>=0.40.0
rich>=13.5.0
gitpython>=3.1.40
pydantic>=2.4.0
typer>=0.9.0
python-dotenv>=1.0.0
tree-sitter>=0.20.1
sentence-transformers>=2.5.0
tiktoken>=0.6.0